+++ src/messages.mli
@@ -40,4 +40,8 @@ val warning : string -> unit
 val ok : string -> unit
 val tips : string -> unit
 
+type answer = Yes | No
+val answer2str : answer -> string
+val confirm : string -> answer
+
 val reset : unit -> unit