+++ src/test/exec_t.ml
@@ -44,23 +44,88 @@ let less_launched test solution () =
   OUnit.assert_equal actual solution
 ;;
 
-(* Data for above test *)
-let ll_data =
-  let max = Const.default_launch in
+(* Function less_launched_num *)
+let less_launched_num test solution () =
+  let actual = Exec_cmd.less_launched_num test in
+  OUnit.assert_equal actual solution
+;;
+
+(* Maximum number of launch *)
+let max = Const.default_launch;;
+(* Data for above test, common data provided to both function since there
+ * purpose are very close from one to the other *)
+let common_data =
   [
-  ( [ ( "cmd1", 4 ) ; ( "cmd2", 0 ) ], Some "cmd2", "Canonical case 1" );
-  ( [ ( "cmd1", 0 ) ; ( "cmd2", 5 ) ], Some "cmd1", "Canonical case 2" );
-  ( [], None, "Empty list" );
-  ( [ ( "cmd1", max ) ; ( "cmd2", (max + 5) ) ], None, "Everything (strcitly) superior to max" );
-  ( [ ( "cmd1", 4 ) ; ( "cmd2", 4 ) ], None, "Twice the same number" );
-]
+    ( [ ( "cmd1", 4 ) ; ( "cmd2", 0 ) ], "Canonical case 1" );
+    ( [ ( "cmd1", 0 ) ; ( "cmd2", 5 ) ], "Canonical case 2" );
+    ( [], "Empty list" );
+    ( [ ( "cmd1", 0 ) ; ( "cmd2", 3 ) ; ( "cmd3", 4 )  ; ( "cmd4", 5 ) ], "Canonical case 3" );
+    ( [ ( "cmd1", 0 ) ; ( "cmd2", 4 ) ; ( "cmd3", 4 )  ; ( "cmd5", 5 ) ],
+      "Twice the same number, with others" );
+    ( [ ( "cmd1", 4 ) ; ( "cmd2", 4 ) ], "Twice the same number" );
+    ( [ ( "cmd1", max ) ; ( "cmd2", (max + 5) ) ],
+      "Everything (strictly) superior to maximum" );
+    (* To prevent >= and > misuse in code *)
+    ( [ ( "cmd1", max - 1 ) ; ( "cmd2", max ) ; ( "cmd3", max + 1 ) ;
+      ( "cmd3", max + 2) ], "Around maximum (ordered)" );
+    ( [ ( "cmd1", max + 1 ) ; ( "cmd2", max ) ; ( "cmd3", max - 1 ) ;
+      ( "cmd3", max + 2) ], "Around maximum (disordered)" )
+  ]
+;;
+(* Add expected result to corresponding to the data provided common set *)
+let add_solutions data expected =
+  List.map2_exn data expected ~f:(fun ( log, name ) solution ->
+    ( log, solution, name ))
+;;
+
+(* Data customized for the tests *)
+let ll_data =
+  add_solutions common_data
+  [
+      Some "cmd2";
+      Some "cmd1";
+      None;
+      Some "cmd1";
+      Some "cmd1";
+      None;
+      None;
+      Some "cmd1";
+      Some "cmd3"
+  ]
+;;
+let ll_data2 =
+  add_solutions common_data
+  [
+    Some 1;
+    Some 0;
+    None;
+    Some 0;
+    Some 0;
+    None;
+    None;
+    Some 0;
+    Some 2
+  ]
+;;
 
 let llt_l =
-  List.map ll_data ~f:(fun (t, s, name) -> ( (less_launched t s), name))
+  let less_launched_suit =
+    List.map ll_data ~f:(fun (t, s, name) -> ( (less_launched t s), name))
+  in
+  less_launched_suit
+  |> List.map ~f:(fun ( f,name ) -> (name, `Quick, f))
+;;
+
+let llt_l2 =
+  let less_launched_num_suit =
+    List.map ll_data2 ~f:(fun (t, s, name) -> ( (less_launched_num t s), name))
+  in
+  less_launched_num_suit
   |> List.map ~f:(fun ( f,name ) -> (name, `Quick, f))
 ;;
 
 (* To be used in test.ml *)
-let alco = [( "Exec_cmd.ml", llt_l );];;
+let alco = [( "Exec_cmd.ml.less_launched", llt_l ) ;
+  ( "Exec_cmd.ml.less_launched_num", llt_l2 )];;